Move options from Builder to Server class
