+++ lib/builder.js
@@ -1,22 +1,14 @@
-const grpc = require('grpc');
-const _ = require('lodash');
-
 module.exports = class {
-  constructor(options) {
-    const defaultOptions = {
-      'host': '0.0.0.0',
-      'port': 3000,
-      'creds': grpc.credentials.createInsecure(),
-    };
-
+  constructor() {
     this._services = [];
     this._middleware = [];
     this._errorHandlers = [];
-    this._options = defaultOptions;
-    _.assign(this._options, options);
   }
 
   addService(protoFilePath, serviceFullName, implementation) {
+    if (!(implementation instanceof Object)) {
+      throw new Error('Cannot add service: No valid implementation received');
+    }
     this._services.push({
       'protoFilePath': protoFilePath,
       'serviceFullName': serviceFullName,
@@ -65,9 +57,5 @@ module.exports = class {
 
   getErrorHandlers() {
     return this._errorHandlers;
-  }
-
-  getOptions() {
-    return this._options;
   }
 };
