+++ lib/builder.spec.js
@@ -1,5 +1,4 @@
 const Builder = require('./builder');
-const grpc = require('grpc');
 
 describe('Builder:', () => {
   let builder, protoFilePath, serviceFullName, implementation, emptyMethod;
@@ -15,6 +14,21 @@ describe('Builder:', () => {
   describe('constructor()', () => {
     it('should create a new instance of Builder class', () => {
       expect(builder instanceof Builder).toBeTruthy();
+    });
+  });
+
+  describe('addService()', () => {
+    describe('implementation is not an object', () => {
+      it('should throw an error', () => {
+        expect(() => {
+          builder.addService(protoFilePath, serviceFullName, 'invalid implementation');
+        }).toThrowError('Cannot add service: No valid implementation received');
+      });
+    });
+
+    it('should return the instance', () => {
+      const instance = builder.addService(protoFilePath, serviceFullName, implementation);
+      expect(instance).toBe(builder);
     });
   });
 
@@ -36,13 +50,6 @@ describe('Builder:', () => {
         builder.addService(protoFilePath, serviceFullName, implementation);
         expect(builder.getServices()).toEqual([expectedService, expectedService]);
       });
-    });
-  });
-
-  describe('addService()', () => {
-    it('should return the instance', () => {
-      const instance = builder.addService(protoFilePath, serviceFullName, implementation);
-      expect(instance).toBe(builder);
     });
   });
 
@@ -129,35 +136,6 @@ describe('Builder:', () => {
         builder.addErrorHandler(scope, emptyMethod);
         builder.addErrorHandler(scope, emptyMethod);
         expect(builder.getErrorHandlers()).toEqual([expectedErrorHandler, expectedErrorHandler]);
-      });
-    });
-  });
-
-  describe('getOptions()', () => {
-    describe('when builder was initialized without options', () => {
-      it('should return the default options', () => {
-        const defaultOptions = {
-          'host': '0.0.0.0',
-          'port': 3000,
-          'creds': grpc.credentials.createInsecure(),
-        };
-        expect(builder.getOptions()).toEqual(defaultOptions);
-      });
-    });
-
-    describe('when builder was initialized with options', () => {
-      beforeEach(() => {
-        const options = {'host': '1.1.1.1'};
-        builder = new Builder(options);
-      });
-
-      it('should return the options merged with the default options', () => {
-        const expectedOptions = {
-          'host': '1.1.1.1',
-          'port': 3000,
-          'creds': grpc.credentials.createInsecure(),
-        };
-        expect(builder.getOptions()).toEqual(expectedOptions);
       });
     });
   });
