+++ lib/server.spec.js
@@ -1,9 +1,10 @@
+const proxyquire = require('proxyquire');
+const grpc = require('grpc');
 const Builder = require('./builder');
 const Mocks = require('../spec/util/mocks');
-const proxyquire = require('proxyquire');
 
 describe('Server:', () => {
-  let Server, server, serverMock, grpcMock, personServiceMock, greeterServiceMock;
+  let Server, server, serverMock, grpcMock, personServiceMock, greeterServiceMock, builder;
 
   beforeEach(() => {
     serverMock = Mocks.getServer();
@@ -11,7 +12,7 @@ describe('Server:', () => {
     personServiceMock = Mocks.getPersonService();
     greeterServiceMock = Mocks.getGreeterService();
     Server = proxyquire('./server', {'grpc': grpcMock});
-    const builder = new Builder();
+    builder = new Builder();
     builder.addService('protoFilePath', 'testapp.PersonService', personServiceMock);
     builder.addService('protoFilePath2', 'testapp.greeter.GreeterService', greeterServiceMock);
     server = new Server(builder);
@@ -20,6 +21,35 @@ describe('Server:', () => {
   describe('constructor()', () => {
     it('should return an instance of the server class', () => {
       expect(server instanceof Server).toBeTruthy();
+    });
+
+    describe('getOptions()', () => {
+      describe('when server was initialized without options', () => {
+        it('should return the default options', () => {
+          const defaultOptions = {
+            'host': '0.0.0.0',
+            'port': 3000,
+            'creds': grpc.credentials.createInsecure(),
+          };
+          expect(server.getOptions()).toEqual(defaultOptions);
+        });
+      });
+
+      describe('when server was initialized with options', () => {
+        beforeEach(() => {
+          const options = {'host': '1.1.1.1'};
+          server = new Server(builder, options);
+        });
+
+        it('should return the options merged with the default options', () => {
+          const expectedOptions = {
+            'host': '1.1.1.1',
+            'port': 3000,
+            'creds': grpc.credentials.createInsecure(),
+          };
+          expect(server.getOptions()).toEqual(expectedOptions);
+        });
+      });
     });
 
     describe('when builder is not set', () => {
